refactor: 네이밍 정리
