+++ src/main/java/step3/LottoBucket.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 import java.util.ArrayList;
 import java.util.Iterator;
