+++ src/main/java/step3/LottoResultView.java
@@ -1,7 +1,6 @@
-package step2;
+package step3;
 
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
