+++ src/main/java/step3/LottoWin.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 import java.util.Optional;
 import java.util.stream.Stream;
