+++ src/main/java/step3/LottoWinStatistics.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 public class LottoWinStatistics {
